+++ src/Core/WorkSpace/src/Tabs/ListTab.tsx
@@ -6,7 +6,7 @@ import { useElementData } from '../../../../packages/Utils/Hooks/useElementData'
 import { WorkspaceFilter } from '../Components/WorkspaceFilter/WorkspaceFilter';
 import { useDataContext } from '../Context/DataProvider';
 import { tabApis } from '../Context/LocationProvider';
-import { useMasterApiContext } from '../Context/TabApiProvider';
+import { useMasterApiContext } from '../Context/MasterApiProvider';
 import { useWorkspaceBookmarks } from '../Util/bookmarks/hooks';
 
 const Wrapper = styled.section`
